fix(login): navbar tablet responsiveness
