+++ src/components/organisms/Navbar.tsx
@@ -15,21 +15,21 @@ export default function Navbar() {
     return (
         <>
             {/* desktop */}
-            <nav className="hidden md:flex md:w-full items-center gap-10 bg-neutral-200 py-6 ">
+            <nav className="hidden lg:flex md:w-full items-center gap-10 justify-center bg-neutral-200 py-6 ">
                 {/* <MdMenu className="aspect-square " size={32} /> */}
-                <div className="flex justify-between max-w-[80vw] mx-auto grow items-center">
-                    <div className="flex gap-5">
-                        <div className="size-10 bg-neutral-400"></div>
-                        <div className="flex gap-2 items-center h-10 px-4 rounded-full bg-neutral-400">
-                            <IoIosSearch size={24} />
+                <div className="flex justify-between gap-5 max-w-[80vw] grow items-center">
+                    <div className="flex gap-5 min-w-0">
+                        <div className="size-10 shrink-0 bg-neutral-400"></div>
+                        <div className="flex shrink min-w-0 gap-2 items-center h-10 px-4 rounded-full bg-neutral-400">
+                            <IoIosSearch size={24} className="shrink-0" />
                             <input
-                                className="h-10 text-neutral-950 outline-none"
+                                className="h-10 min-w-0 shrink text-neutral-950 outline-none"
                                 type="text"
                                 placeholder="Search"
                             />
                         </div>
                     </div>
-                    <div className="flex items-center flex gap-[25px]">
+                    <div className="flex items-center whitespace-nowrap text-md flex gap-[25px]">
                         <p>Nav Link</p>
                         <p>Nav Link</p>
                         <p>Nav Link</p>
@@ -44,7 +44,7 @@ export default function Navbar() {
             </nav>
             {/* mobile */}
             <Sheet>
-                <nav className="flex md:hidden md:w-full items-center gap-10 bg-neutral-200 py-6 ">
+                <nav className="flex lg:hidden md:w-full items-center gap-10 bg-neutral-200 py-6 ">
                     {/* <MdMenu className="aspect-square " size={32} /> */}
                     <div className="flex justify-between max-w-[80vw] mx-auto grow items-center">
                         <div className="flex gap-5">
